+++ packages/frontend/src/components/DashboardFilter/FilterConfiguration/FilterSettings.tsx
@@ -1,4 +1,3 @@
-import { HTMLSelect } from '@blueprintjs/core';
 import { Popover2Props } from '@blueprintjs/popover2';
 import {
     DashboardFilterRule,
@@ -8,7 +7,7 @@ import {
     getFilterRuleWithDefaultValue,
     getFilterTypeFromItem,
 } from '@lightdash/common';
-import { Stack, Switch, TextInput, Tooltip } from '@mantine/core';
+import { Select, Stack, Switch, TextInput, Tooltip } from '@mantine/core';
 import { FC, useEffect, useMemo, useState } from 'react';
 import { FilterTypeConfig } from '../../common/Filters/configs';
 import { getPlaceholderByFilterTypeAndOperator } from '../../common/Filters/utils/getPlaceholderByFilterTypeAndOperator';
@@ -30,9 +29,7 @@ const FilterSettings: FC<FilterSettingsProps> = ({
     onChangeFilterOperator,
     onChangeFilterRule,
 }) => {
-    const [filterLabel, setFilterLabel] = useState<string>(
-        filterRule.label ?? field.label,
-    );
+    const [filterLabel, setFilterLabel] = useState<string>();
     const filterType = field ? getFilterTypeFromItem(field) : FilterType.STRING;
 
     const filterConfig = useMemo(
@@ -50,6 +47,13 @@ const FilterSettings: FC<FilterSettingsProps> = ({
             });
         }
     }, [isEditMode, onChangeFilterRule, filterRule]);
+
+    // Set default label when using revert (undo) button
+    useEffect(() => {
+        if (filterLabel !== '') {
+            setFilterLabel(filterRule.label ?? field.label);
+        }
+    }, [filterLabel, filterRule.label, field.label]);
 
     return (
         <Stack>
@@ -90,14 +94,12 @@ const FilterSettings: FC<FilterSettingsProps> = ({
                     </Tooltip>
                 )}
 
-                <HTMLSelect
-                    fill
-                    onChange={(e) =>
-                        onChangeFilterOperator(
-                            e.target.value as FilterRule['operator'],
-                        )
+                <Select
+                    size="xs"
+                    data={filterConfig.operatorOptions}
+                    onChange={(operator: FilterRule['operator']) =>
+                        onChangeFilterOperator(operator)
                     }
-                    options={filterConfig.operatorOptions}
                     value={filterRule.operator}
                 />
                 {filterRule.disabled ? (